+++ lib/plivo/xml/response.rb
@@ -2,7 +2,7 @@ module Plivo
   module XML
     class Response < Element
       @nestables = %w[Speak Play GetDigits GetInput Record Dial Message
-                      Redirect Wait Hangup PreAnswer Conference DTMF MultiPartyCall]
+                      Redirect Wait Hangup PreAnswer Conference DTMF MultiPartyCall Stream]
       @valid_attributes = []
 
       def initialize